Create module graph (from utils.get_dag and utils.get_filter)
It's named `graph`, not `dag`, because we use `dag` as variable name
in too many places.
